+++ vault/resource_aws_secret_backend.go
@@ -6,6 +6,7 @@ package vault
 import (
 	"context"
 	"fmt"
+	automatedrotationutil "github.com/hashicorp/terraform-provider-vault/internal/rotation"
 	"log"
 	"strings"
 
@@ -163,6 +164,9 @@ func awsSecretBackendResource() *schema.Resource {
 		consts.FieldLocal,
 	))
 
+	// Add common automated root rotation schema to the resource
+	provider.MustAddSchema(r, provider.GetAutomatedRootRotationSchema())
+
 	return r
 }
 
@@ -193,7 +197,8 @@ func getMountCustomizeDiffFunc(field string) schema.CustomizeDiffFunc {
 
 func awsSecretBackendCreate(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	useAPIVer119 := provider.IsAPISupported(meta, provider.VaultVersion119)
-	useAPIVer116 := provider.IsAPISupported(meta, provider.VaultVersion116) && provider.IsEnterpriseSupported(meta)
+	isEnterprise := provider.IsEnterpriseSupported(meta)
+	useAPIVer116Enterprise := provider.IsAPISupported(meta, provider.VaultVersion116) && isEnterprise
 
 	client, e := provider.GetClient(d, meta)
 	if e != nil {
@@ -239,9 +244,14 @@ func awsSecretBackendCreate(ctx context.Context, d *schema.ResourceData, meta in
 		if v, ok := d.GetOk(consts.FieldSTSRegion); ok {
 			data[consts.FieldSTSRegion] = v.(string)
 		}
+
+		// parse automated root rotation fields if Enterprise 1.19 server
+		if isEnterprise {
+			automatedrotationutil.ParseAutomatedRotationFields(d, data)
+		}
 	}
 
-	if useAPIVer116 {
+	if useAPIVer116Enterprise {
 		if v, ok := d.GetOk(consts.FieldIdentityTokenAudience); ok && v != "" {
 			data[consts.FieldIdentityTokenAudience] = v.(string)
 		}
@@ -271,7 +281,8 @@ func awsSecretBackendCreate(ctx context.Context, d *schema.ResourceData, meta in
 }
 
 func awsSecretBackendRead(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
-	useAPIVer116 := provider.IsAPISupported(meta, provider.VaultVersion116) && provider.IsEnterpriseSupported(meta)
+	isEnterprise := provider.IsEnterpriseSupported(meta)
+	useAPIVer116 := provider.IsAPISupported(meta, provider.VaultVersion116) && isEnterprise
 	useAPIVer119 := provider.IsAPISupported(meta, provider.VaultVersion119)
 
 	client, e := provider.GetClient(d, meta)
@@ -337,6 +348,12 @@ func awsSecretBackendRead(ctx context.Context, d *schema.ResourceData, meta inte
 					return diag.Errorf("error reading %s for AWS Secret Backend %q: %q", consts.FieldSTSRegion, path, err)
 				}
 			}
+
+			if isEnterprise {
+				if err := automatedrotationutil.PopulateAutomatedRotationFields(d, resp, path); err != nil {
+					return diag.FromErr(err)
+				}
+			}
 		}
 
 		if useAPIVer116 {
@@ -364,7 +381,8 @@ func awsSecretBackendRead(ctx context.Context, d *schema.ResourceData, meta inte
 }
 
 func awsSecretBackendUpdate(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
-	useAPIVer116 := provider.IsAPISupported(meta, provider.VaultVersion116) && provider.IsEnterpriseSupported(meta)
+	isEnterprise := provider.IsEnterpriseSupported(meta)
+	useAPIVer116 := provider.IsAPISupported(meta, provider.VaultVersion116) && isEnterprise
 	useAPIVer119 := provider.IsAPISupported(meta, provider.VaultVersion119)
 
 	client, e := provider.GetClient(d, meta)
@@ -386,6 +404,10 @@ func awsSecretBackendUpdate(ctx context.Context, d *schema.ResourceData, meta in
 		consts.FieldSecretKey, consts.FieldRegion, consts.FieldIAMEndpoint,
 		consts.FieldSTSEndpoint, consts.FieldSTSFallbackEndpoints, consts.FieldSTSRegion, consts.FieldSTSFallbackRegions,
 		consts.FieldIdentityTokenTTL, consts.FieldIdentityTokenAudience, consts.FieldRoleArn,
+		consts.FieldRotationSchedule,
+		consts.FieldRotationPeriod,
+		consts.FieldRotationWindow,
+		consts.FieldDisableAutomatedRotation,
 	) {
 		log.Printf("[DEBUG] Updating root credentials at %q", path+"/config/root")
 		data := map[string]interface{}{
@@ -410,6 +432,11 @@ func awsSecretBackendUpdate(ctx context.Context, d *schema.ResourceData, meta in
 
 			if v, ok := d.GetOk(consts.FieldSTSRegion); ok {
 				data[consts.FieldSTSRegion] = v.(string)
+			}
+
+			// parse automated root rotation fields if Enterprise 1.19 server
+			if isEnterprise {
+				automatedrotationutil.ParseAutomatedRotationFields(d, data)
 			}
 		}